// types/generated/api.d.ts
export interface components {
    schemas: {
// ...
        RoleDto: {
            /** @example role-author */
            id: string;
            /** @enum {string} */
            name: "Author" | "Reviewer" | "Admin";
            permissions: components["schemas"]["PermissionDto"][];
        };
// ...
        LoginResponseDto: {
            /** @example eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9... */
            accessToken: string;
            user: components["schemas"]["UserDto"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}